refactor: change components placement
- rename dashboard file to make it easier to import
- create new file for all styled components used in dashboard
